reimplement Fn as 'aux' rather than gamepad button
same general idea, but we don't reserve a gamepad mask for something
that will never be used as a gamepad button this way. this also
resurrects the "settings pin" idea and the remnants of its aux
mechanism, while still allowing it to be a web configurable pin. it is
still outwardly called the Fn or Function button, but it is assigned to
a separate gamepad state and thus is a bit more obvious that it is *not*
a button that will get sent to the host